add class PredicateExtractor
